+++ Tests/BugsnagClientMirrorTest.m
@@ -95,7 +95,8 @@
             @"context @16@0:8",
             @"collectAppWithState @16@0:8",
             @"collectBreadcrumbs @16@0:8",
-            @"collectThreads @16@0:8",
+            @"collectThreads: @20@0:8B16",
+            @"collectThreads: @20@0:8c16",
             @"collectDeviceWithState @16@0:8",
             @"extraRuntimeInfo @16@0:8",
             @"setExtraRuntimeInfo: v24@0:8@16",
@@ -116,7 +117,8 @@
             @"appendNSErrorInfo:block:event: c40@0:8@16@?24@32",
             @"createNSErrorWrapper: @24@0:8@16",
             @"setBreadcrumbs: v24@0:8@16",
-            @"breadcrumbs @16@0:8"
+            @"breadcrumbs @16@0:8",
+            @"setUser: v24@0:8@16"
     ]];
 
     // the following methods are implemented on Bugsnag but do not need to